+++ changelog.md
@@ -86,6 +86,7 @@ Codion Change Log
 - FontImageIcon.Builder.IkonBuilder added.
 - TextFieldBuilder.ValueClassBuilder added.
 - TemporalField.Builder.TemporalClassBuilder added.
+- TemporalFieldPanel.Builder.TemporalClassBuilder added.
 ### is.codion.swing.framework.model
 - EntityComboBoxModel.Builder.EntityTypeBuilder and ConnectionProviderBuilder added.
 ### is.codion.swing.framework.ui